pyproject.toml
[project]
name = "knx-nats-bridge"
version = "0.1.0"
description = "Bridge KNX telegrams to NATS JetStream with DPT decoding"
readme = "README.md"
license = "GPL-2.0-or-later"
requires-python = ">=3.12"
authors = [
    { name = "Alexander Zimmermann" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU General Public License v2 or later (GPLv2+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
]
dependencies = [
    "xknx>=3.5.0",
    "nats-py>=2.8.0",
    "prometheus-client>=0.20.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.1",
    "jsonschema>=4.22.0",
]

[project.optional-dependencies]
tools = [
    "xknxproject>=3.8.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "types-PyYAML",
    "types-jsonschema",
]

[project.scripts]
knx-nats-bridge = "knx_nats_bridge.main:run"
knxproj-to-yaml = "knx_nats_bridge.tools.knxproj_to_yaml:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/knx_nats_bridge"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "RET", "SIM", "ARG"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
